+++ CLASSES/old_stuff.txt
@@ -1,0 +1,7 @@
+# self.features = {}
+        # with open(f"./TXT/{name}.txt", 'r') as file:
+        #     reader = csv.reader(file)
+        #     for row in reader:
+        #         if row:
+        #             key, feature = row
+        #             self.features[key] = feature
