+++ TXT/Old Beggar.txt
@@ -1,0 +1,11 @@
+first_name,Old
+last_name,Beggar
+occupation,Cowboy
+age,55
+description,"Old Beggar is an old beggar. He believes he's a Hollywood actor. He's got some stuff for sale."
+stats,0 0 0
+items,piece_of_rope broken_cell_phone
+skills,begging sleeping
+talk,"Back in the day I was a famous actor. I had ladies by the dozen and as much drugs as you can possibly imagine. But then I trusted this skunk and all went to shit".
+
+
