+++ TXT/blade.txt
@@ -1,0 +1,3 @@
+name,blade
+description,"A low class knife, barely sufficient for self-defense against children and old ladies."
+stats,10 0 0 0 100 500
