+++ TXT/crossroads.txt
