+++ TXT/roamer.txt
@@ -1,0 +1,10 @@
+first_name,None
+last_name,None
+age,0
+occupation,roamer
+bio,"Roamers are nothing short of scavengers. They take what they can get."
+description,"You're a tall, skinny, sketchy fucker with murder in your eyes and darkness in your heart."
+items,blade sunglasses
+skills,evade
+stats,60 40 80
+
